KIALI-1129 Show virtuals services in a table within service details page (#510)
* KIALI-1129 Show virtuals services in a table within service details page

* KIALI-1129 Adding warning/issues count into icon infotip

* KIALI-1129 Adjusting columns and links
