+++ src/pages/ServiceDetails/ServiceInfo/ServiceInfoVirtualServices.scss
@@ -1,0 +1,9 @@
+table.table {
+  .status {
+    width: 50px;
+  }
+
+  .actions {
+    width: 85px;
+  }
+}
